+++ src/lib/rsc-handler-worker.ts
@@ -7,16 +7,21 @@ import { createServer } from "vite";
 import { createElement } from "react";
 import RSDWServer from "react-server-dom-webpack/server";
 
+import { configFileConfig, resolveConfig } from "./config.js";
 import { transformRsfId, generatePrefetchCode } from "./rsc-utils.js";
-import type { RenderInput, MessageReq, MessageRes } from "./rsc-handler.js";
-import type { Config } from "../../config.js";
-import { defineEntries } from "../../server.js";
-import type { unstable_GetCustomModules } from "../../server.js";
+import type {
+  RenderInput,
+  MessageReq,
+  MessageRes,
+  BuildOutput,
+} from "./rsc-handler.js";
+import { defineEntries } from "../server.js";
 import { rscTransformPlugin, rscReloadPlugin } from "./vite-plugin-rsc.js";
 
 const { renderToPipeableStream } = RSDWServer;
 
 type Entries = { default: ReturnType<typeof defineEntries> };
+type PipeableStream = { pipe<T extends Writable>(destination: T): T };
 
 const handleSetClientEntries = async (
   mesg: MessageReq & { type: "setClientEntries" }
@@ -24,17 +29,10 @@ const handleSetClientEntries = async (
   const { id, value } = mesg;
   try {
     await setClientEntries(value);
-    const mesg: MessageRes = {
-      id,
-      type: "end",
-    };
+    const mesg: MessageRes = { id, type: "end" };
     parentPort!.postMessage(mesg);
   } catch (err) {
-    const mesg: MessageRes = {
-      id,
-      type: "err",
-      err,
-    };
+    const mesg: MessageRes = { id, type: "err", err };
     parentPort!.postMessage(mesg);
   }
 };
@@ -60,43 +58,14 @@ const handleRender = async (mesg: MessageReq & { type: "render" }) => {
         callback();
       },
       final(callback) {
-        const mesg: MessageRes = {
-          id,
-          type: "end",
-        };
+        const mesg: MessageRes = { id, type: "end" };
         parentPort!.postMessage(mesg);
         callback();
       },
     });
     pipeable.pipe(writable);
   } catch (err) {
-    const mesg: MessageRes = {
-      id,
-      type: "err",
-      err,
-    };
-    parentPort!.postMessage(mesg);
-  }
-};
-
-const handleGetCustomModules = async (
-  mesg: MessageReq & { type: "getCustomModules" }
-) => {
-  const { id } = mesg;
-  try {
-    const modules = await getCustomModulesRSC();
-    const mesg: MessageRes = {
-      id,
-      type: "customModules",
-      modules,
-    };
-    parentPort!.postMessage(mesg);
-  } catch (err) {
-    const mesg: MessageRes = {
-      id,
-      type: "err",
-      err,
-    };
+    const mesg: MessageRes = { id, type: "err", err };
     parentPort!.postMessage(mesg);
   }
 };
@@ -104,78 +73,23 @@ const handleGetCustomModules = async (
 const handleBuild = async (mesg: MessageReq & { type: "build" }) => {
   const { id } = mesg;
   try {
-    await buildRSC();
-    const mesg: MessageRes = {
-      id,
-      type: "end",
-    };
+    const output = await buildRSC();
+    const mesg: MessageRes = { id, type: "buildOutput", output };
     parentPort!.postMessage(mesg);
   } catch (err) {
-    const mesg: MessageRes = {
-      id,
-      type: "err",
-      err,
-    };
+    const mesg: MessageRes = { id, type: "err", err };
     parentPort!.postMessage(mesg);
   }
 };
 
-parentPort!.on("message", (mesg: MessageReq) => {
-  if (mesg.type === "shutdown") {
-    vitePromise.then(async (vite) => {
-      await vite.close();
-      parentPort!.close();
-    });
-  } else if (mesg.type === "setClientEntries") {
-    handleSetClientEntries(mesg);
-  } else if (mesg.type === "render") {
-    handleRender(mesg);
-  } else if (mesg.type === "getCustomModules") {
-    handleGetCustomModules(mesg);
-  } else if (mesg.type === "build") {
-    handleBuild(mesg);
-  }
-});
-
-type PipeableStream = {
-  pipe<T extends Writable>(destination: T): T;
-};
-
-// TODO use of process.env is all temporary
-// TODO these are temporary
-const config: Config =
-  (process.env.WAKU_CONFIG && JSON.parse(process.env.WAKU_CONFIG)) || {};
-const dirFromConfig =
-  config.prdServer?.dir ?? config.build?.dir ?? config.devServer?.dir; // HACK
-const dir = path.resolve(dirFromConfig || ".");
-const basePath = config.build?.basePath || "/"; // FIXME it's not build only
-const distPath = config.files?.dist || "dist";
-const publicPath = path.join(distPath, config.files?.public || "public");
-const publicIndexHtmlFile = path.join(
-  dir,
-  publicPath,
-  config.files?.indexHtml || "index.html"
-);
-const entriesFile = path.join(dir, config.files?.entriesJs || "entries.js");
-const distEntriesFile = path.join(
-  dir,
-  distPath,
-  config.files?.entriesJs || "entries.js"
-);
-
 const vitePromise = createServer({
-  root: dir,
-  ...(process.env.NODE_ENV && { mode: process.env.NODE_ENV }),
+  ...configFileConfig,
   plugins: [
     rscTransformPlugin(),
-    ...(process.env.NODE_ENV === "development"
-      ? [
-          rscReloadPlugin((type) => {
-            const mesg: MessageRes = { type };
-            parentPort!.postMessage(mesg);
-          }),
-        ]
-      : []),
+    rscReloadPlugin((type) => {
+      const mesg: MessageRes = { type };
+      parentPort!.postMessage(mesg);
+    }),
   ],
   ssr: {
     // FIXME Without this, "use client" directive in waku/router/client
@@ -187,17 +101,54 @@ const vitePromise = createServer({
   appType: "custom",
 });
 
+const shutdown = async () => {
+  const vite = await vitePromise;
+  await vite.close();
+  parentPort!.close();
+};
+
 const loadServerFile = async (fname: string) => {
   const vite = await vitePromise;
   return vite.ssrLoadModule(fname);
 };
 
-const getFunctionComponent = async (rscId: string, isBuild: boolean) => {
+parentPort!.on("message", (mesg: MessageReq) => {
+  if (mesg.type === "shutdown") {
+    shutdown();
+  } else if (mesg.type === "setClientEntries") {
+    handleSetClientEntries(mesg);
+  } else if (mesg.type === "render") {
+    handleRender(mesg);
+  } else if (mesg.type === "build") {
+    handleBuild(mesg);
+  }
+});
+
+const configPromise = resolveConfig("serve");
+
+const getEntriesFile = async (
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  isBuild: boolean
+) => {
+  if (isBuild) {
+    return path.join(
+      config.root,
+      config.build.outDir,
+      config.framework.entriesJs
+    );
+  }
+  return path.join(config.root, config.framework.entriesJs);
+};
+
+const getFunctionComponent = async (
+  rscId: string,
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  isBuild: boolean
+) => {
+  const entriesFile = await getEntriesFile(config, isBuild);
   const {
     default: { getEntry },
-  } = await (loadServerFile(
-    isBuild ? distEntriesFile : entriesFile
-  ) as Promise<Entries>);
+  } = await (loadServerFile(entriesFile) as Promise<Entries>);
   const mod = await getEntry(rscId);
   if (typeof mod === "function") {
     return mod;
@@ -210,11 +161,14 @@ const getFunctionComponent = async (rscId: string, isBuild: boolean) => {
 
 let absoluteClientEntries: Record<string, string> = {};
 
-const resolveClientEntry = (filePath: string) => {
+const resolveClientEntry = (
+  config: Awaited<ReturnType<typeof resolveConfig>>,
+  filePath: string
+) => {
   const clientEntry = absoluteClientEntries[filePath];
   if (!clientEntry) {
     if (absoluteClientEntries["*"] === "*") {
-      return basePath + path.relative(dir, filePath);
+      return config.base + path.relative(config.root, filePath);
     }
     throw new Error("No client entry found for " + filePath);
   }
@@ -228,6 +182,8 @@ async function setClientEntries(
     absoluteClientEntries = value;
     return;
   }
+  const config = await configPromise;
+  const entriesFile = await getEntriesFile(config, false);
   const { clientEntries } = await loadServerFile(entriesFile);
   if (!clientEntries) {
     throw new Error("Failed to load clientEntries");
@@ -236,18 +192,19 @@ async function setClientEntries(
   absoluteClientEntries = Object.fromEntries(
     Object.entries(clientEntries).map(([key, val]) => [
       path.join(baseDir, key),
-      basePath + val,
+      config.base + val,
     ])
   );
 }
 
 async function renderRSC(input: RenderInput): Promise<PipeableStream> {
+  const config = await configPromise;
   const bundlerConfig = new Proxy(
     {},
     {
       get(_target, encodedId: string) {
         const [filePath, name] = encodedId.split("#") as [string, string];
-        const id = resolveClientEntry(filePath);
+        const id = resolveClientEntry(config, filePath);
         return { id, chunks: [id], name, async: true };
       },
     }
@@ -255,7 +212,7 @@ async function renderRSC(input: RenderInput): Promise<PipeableStream> {
 
   if (input.rsfId && input.args) {
     const [fileId, name] = input.rsfId.split("#");
-    const fname = path.join(dir, fileId!);
+    const fname = path.join(config.root, fileId!);
     const mod = await loadServerFile(fname);
     const data = await (mod[name!] || mod)(...input.args);
     if (!input.rscId) {
@@ -264,32 +221,20 @@ async function renderRSC(input: RenderInput): Promise<PipeableStream> {
     // continue for mutation mode
   }
   if (input.rscId && input.props) {
-    const component = await getFunctionComponent(input.rscId, false);
+    const component = await getFunctionComponent(input.rscId, config, false);
     return renderToPipeableStream(
       createElement(component, input.props),
       bundlerConfig
-    ).pipe(transformRsfId(dir));
+    ).pipe(transformRsfId(config.root));
   }
   throw new Error("Unexpected input");
 }
 
-async function getCustomModulesRSC(): Promise<{ [name: string]: string }> {
-  const {
-    default: { unstable_getCustomModules: getCustomModules },
-  } = await (loadServerFile(entriesFile) as Promise<{
-    default: Entries["default"] & {
-      unstable_getCustomModules?: unstable_GetCustomModules;
-    };
-  }>);
-  if (!getCustomModules) {
-    return {};
-  }
-  const modules = await getCustomModules();
-  return modules;
-}
-
 // FIXME this may take too much responsibility
-async function buildRSC(): Promise<void> {
+async function buildRSC(): Promise<BuildOutput> {
+  const config = await resolveConfig("build");
+  const basePrefix = config.base + config.framework.rscPrefix;
+  const distEntriesFile = await getEntriesFile(config, true);
   const {
     default: { getBuilder },
   } = await (loadServerFile(distEntriesFile) as Promise<Entries>);
@@ -297,17 +242,31 @@ async function buildRSC(): Promise<void> {
     console.warn(
       "getBuilder is undefined. It's recommended for optimization and sometimes required."
     );
-    return;
+    return { rscFiles: [], htmlFiles: [] };
   }
 
-  // FIXME this doesn't seem an ideal solution
-  const decodeId = (encodedId: string): [id: string, name: string] => {
-    const [filePath, name] = encodedId.split("#") as [string, string];
-    const id = resolveClientEntry(filePath);
-    return [id, name];
+  const renderForBuild = (
+    element: unknown,
+    clientModuleCallback: (id: string) => void
+  ): PipeableStream => {
+    const bundlerConfig = new Proxy(
+      {},
+      {
+        get(_target, encodedId: string) {
+          const [filePath, name] = encodedId.split("#") as [string, string];
+          const id = resolveClientEntry(config, filePath);
+          clientModuleCallback(id);
+          return { id, chunks: [id], name, async: true };
+        },
+      }
+    );
+    const pipeable = renderToPipeableStream(element, bundlerConfig).pipe(
+      transformRsfId(path.join(config.root, config.build.outDir))
+    );
+    return pipeable;
   };
 
-  const pathMap = await getBuilder(decodeId);
+  const pathMap = await getBuilder(config.root, renderForBuild);
   const clientModuleMap = new Map<string, Set<string>>();
   const addClientModule = (pathStr: string, id: string) => {
     let idSet = clientModuleMap.get(pathStr);
@@ -317,6 +276,7 @@ async function buildRSC(): Promise<void> {
     }
     idSet.add(id);
   };
+  const rscFileSet = new Set<string>(); // XXX could be implemented better
   await Promise.all(
     Object.entries(pathMap).map(async ([pathStr, { elements }]) => {
       for (const [rscId, props] of elements || []) {
@@ -325,46 +285,46 @@ async function buildRSC(): Promise<void> {
         const searchParams = new URLSearchParams();
         searchParams.set("props", serializedProps);
         const destFile = path.join(
-          dir,
-          publicPath,
-          "RSC",
-          decodeURIComponent(rscId),
+          config.root,
+          config.build.outDir,
+          config.framework.outPublic,
+          config.framework.rscPrefix + decodeURIComponent(rscId),
           decodeURIComponent(`${searchParams}`)
         );
-        fs.mkdirSync(path.dirname(destFile), { recursive: true });
-        const bundlerConfig = new Proxy(
-          {},
-          {
-            get(_target, encodedId: string) {
-              const [id, name] = decodeId(encodedId);
-              addClientModule(pathStr, id);
-              return { id, chunks: [id], name, async: true };
-            },
-          }
-        );
-        const component = await getFunctionComponent(rscId, true);
-        const pipeable = renderToPipeableStream(
-          createElement(component, props as any),
-          bundlerConfig
-        ).pipe(transformRsfId(path.join(dir, distPath)));
-        await new Promise<void>((resolve, reject) => {
-          const stream = fs.createWriteStream(destFile);
-          stream.on("finish", resolve);
-          stream.on("error", reject);
-          pipeable.pipe(stream);
-        });
+        if (!rscFileSet.has(destFile)) {
+          rscFileSet.add(destFile);
+          fs.mkdirSync(path.dirname(destFile), { recursive: true });
+          const component = await getFunctionComponent(rscId, config, true);
+          const pipeable = renderForBuild(
+            createElement(component, props as any),
+            (id) => addClientModule(pathStr, id)
+          );
+          await new Promise<void>((resolve, reject) => {
+            const stream = fs.createWriteStream(destFile);
+            stream.on("finish", resolve);
+            stream.on("error", reject);
+            pipeable.pipe(stream);
+          });
+        }
       }
     })
   );
 
+  const publicIndexHtmlFile = path.join(
+    config.root,
+    config.build.outDir,
+    config.framework.outPublic,
+    config.framework.indexHtml
+  );
   const publicIndexHtml = fs.readFileSync(publicIndexHtmlFile, {
     encoding: "utf8",
   });
-  await Promise.all(
+  const htmlFiles = await Promise.all(
     Object.entries(pathMap).map(async ([pathStr, { elements, customCode }]) => {
       const destFile = path.join(
-        dir,
-        publicPath,
+        config.root,
+        config.build.outDir,
+        config.framework.outPublic,
         pathStr,
         pathStr.endsWith("/") ? "index.html" : ""
       );
@@ -377,6 +337,7 @@ async function buildRSC(): Promise<void> {
       }
       const code =
         generatePrefetchCode(
+          basePrefix,
           Array.from(elements || []).flatMap(([rscId, props, skipPrefetch]) => {
             if (skipPrefetch) {
               return [];
@@ -390,6 +351,8 @@ async function buildRSC(): Promise<void> {
         data = data.replace(/<\/body>/, `<script>${code}</script></body>`);
       }
       fs.writeFileSync(destFile, data, { encoding: "utf8" });
+      return destFile;
     })
   );
+  return { rscFiles: Array.from(rscFileSet), htmlFiles };
 }