+++ src/lib/rsc-handler.ts
@@ -13,8 +13,9 @@ export type RenderInput<Props extends {} = {}> = {
   args?: unknown[] | undefined;
 };
 
-type CustomModules = {
-  [name: string]: string;
+export type BuildOutput = {
+  rscFiles: string[];
+  htmlFiles: string[];
 };
 
 export type MessageReq =
@@ -24,26 +25,15 @@ export type MessageReq =
       type: "setClientEntries";
       value: "load" | Record<string, string>;
     }
-  | {
-      id: number;
-      type: "render";
-      input: RenderInput;
-    }
-  | {
-      id: number;
-      type: "getCustomModules";
-    }
-  | {
-      id: number;
-      type: "build";
-    };
+  | { id: number; type: "render"; input: RenderInput }
+  | { id: number; type: "build" };
 
 export type MessageRes =
   | { type: "full-reload" }
   | { id: number; type: "buf"; buf: ArrayBuffer; offset: number; len: number }
   | { id: number; type: "end" }
   | { id: number; type: "err"; err: unknown }
-  | { id: number; type: "customModules"; modules: CustomModules };
+  | { id: number; type: "buildOutput"; output: BuildOutput };
 
 const messageCallbacks = new Map<number, (mesg: MessageRes) => void>();
 
@@ -66,7 +56,8 @@ export function registerReloadCallback(fn: (type: "full-reload") => void) {
 export function shutdown() {
   return new Promise<void>((resolve) => {
     worker.on("close", resolve);
-    worker.postMessage({ type: "shutdown" });
+    const mesg: MessageReq = { type: "shutdown" };
+    worker.postMessage(mesg);
   });
 }
 
@@ -86,11 +77,7 @@ export function setClientEntries(
         messageCallbacks.delete(id);
       }
     });
-    const mesg: MessageReq = {
-      id,
-      type: "setClientEntries",
-      value,
-    };
+    const mesg: MessageReq = { id, type: "setClientEntries", value };
     worker.postMessage(mesg);
   });
 }
@@ -111,51 +98,24 @@ export function renderRSC(input: RenderInput): Readable {
       messageCallbacks.delete(id);
     }
   });
-  const mesg: MessageReq = {
-    id,
-    type: "render",
-    input,
-  };
+  const mesg: MessageReq = { id, type: "render", input };
   worker.postMessage(mesg);
   return passthrough;
 }
 
-export function getCustomModulesRSC(): Promise<CustomModules> {
-  return new Promise<CustomModules>((resolve, reject) => {
+export function buildRSC(): Promise<BuildOutput> {
+  return new Promise((resolve, reject) => {
     const id = nextId++;
     messageCallbacks.set(id, (mesg) => {
-      if (mesg.type === "customModules") {
-        resolve(mesg.modules);
+      if (mesg.type === "buildOutput") {
+        resolve(mesg.output);
         messageCallbacks.delete(id);
       } else if (mesg.type === "err") {
         reject(mesg.err);
         messageCallbacks.delete(id);
       }
     });
-    const mesg: MessageReq = {
-      id,
-      type: "getCustomModules",
-    };
-    worker.postMessage(mesg);
-  });
-}
-
-export function buildRSC(): Promise<void> {
-  return new Promise<void>((resolve, reject) => {
-    const id = nextId++;
-    messageCallbacks.set(id, (mesg) => {
-      if (mesg.type === "end") {
-        resolve();
-        messageCallbacks.delete(id);
-      } else if (mesg.type === "err") {
-        reject(mesg.err);
-        messageCallbacks.delete(id);
-      }
-    });
-    const mesg: MessageReq = {
-      id,
-      type: "build",
-    };
+    const mesg: MessageReq = { id, type: "build" };
     worker.postMessage(mesg);
   });
 }